// src/app/core/model/quski/TbQoRolTipoDocumento.ts
export class TbQoRolTipoDocumento {
    id: number;
    idTipoDocumento: number;
    idRol: number;
    estado: string;
    proceso: string;
    estadoOperacion: string;
    descargaPlantilla: boolean;
    cargaDocumento: boolean;
    descargaDocumento: boolean;
    cargaDocumentoAdicional: boolean;
    descargaDocumentoAdicional: boolean;
    accionUno: boolean;
    accionDos: boolean;
    accionTres: boolean;
}